implemented permission handling code-cleanup some minor bug-fixes

+++ app/controllers/time_bookings_controller.rb
@@ -29,7 +29,7 @@ class TimeBookingsController < ApplicationController
     issue = Issue.where(:id => tb[:issue_id]).first
     issue.nil? ? project = Project.where(:id => tb[:project_id]).first : project = issue.project
 
-    if time_booking.user.id == User.current.id || User.current.admin?
+    if time_booking.user.id == User.current.id && User.current.allowed_to_globally?(:tt_edit_own_time_bookings, {}) || User.current.allowed_to_globally?(:tt_edit_time_bookings, {})
       start = Time.parse(tb[:tt_booking_date] + " " + tb[:start_time])
       hours = time_string2hour(tb[:spent_time])
       stop = start + hours.hours
@@ -42,20 +42,37 @@ class TimeBookingsController < ApplicationController
 
       time_booking.save!
       tl.check_bookable
+      flash[:notice] = l(:tt_update_booking_success)
+    else
+      flash[:error] = l(:tt_update_booking_not_allowed)
     end
+  rescue TimeBookingError => e
+    flash[:error] = e.message
   end
 
   def delete
-    time_bookings = TimeBooking.where(:id => params[:time_booking_ids]).all
-    time_bookings.each do |item|
-      tl = TimeLog.where(:id => item.time_log_id, :user_id => User.current.id).first
-      item.destroy
-      tl.check_bookable # we should set the bookable_flag after deleting bookings
+    if help.permission_checker([:tt_edit_own_time_bookings, :tt_edit_time_bookings], {}, true)
+      time_bookings = TimeBooking.where(:id => params[:time_booking_ids]).all
+      time_bookings.each do |item|
+        if item.user == User.current && User.current.allowed_to_globally?(:tt_edit_own_time_bookings, {}) || User.current.allowed_to_globally?(:tt_edit_time_bookings, {})
+          tl = TimeLog.where(:id => item.time_log_id, :user_id => User.current.id).first
+          item.destroy
+          tl.check_bookable # we should set the bookable_flag after deleting bookings
+        else
+          flash[:error] = l(:tt_error_not_allowed_to_delete_bookings)
+        end
+      end
       flash[:notice] = l(:time_tracker_delete_booking_success)
+    else
+      flash[:error] = l(:tt_error_not_allowed_to_delete_bookings)
     end
+    redirect_to :back
+  rescue TimeBookingError => e
+    flash[:error] = e.message
     redirect_to :back
   end
 
+  # TODO check if there should be done TimeBooking.visible.where....
   def get_list_entry
     # prepare query for time_bookings
     time_bookings_query
